New item - Mysterious Painting, added NBT info to ItemCustomPainting
ItemRandomPainting - no recipe yet, when used gives a random custom
painting to the player. Have a chance to give an authentic version or a
copy.

Authentic painting have a 'RARE' item rarity, but no other function.

Also started to add NBT support for the ItemCustomPainting, the canvas
ID will be controlled via NBT, not damage values.